added delete project dialog

+++ app/src/redux/middleware/database-middleware.js
@@ -4,8 +4,8 @@ import ActionTypes from '../actions/action-types'
 
 export default database => {
     return ({dispatch}) => {
-       
-        database.onProjects = projects => 
+
+        database.onProjects = projects =>
             dispatch(ActionsGenerator.onProjects(projects))
 
         return next => action => {
@@ -20,8 +20,9 @@ export default database => {
                     database.submitProject(action.name, action.description)
                     break
 
-                case ActionTypes.DELETE_PROJECT:
-                    database.deleteProject(action.uid)
+                case ActionTypes.DELETE_PROJECT_CONFIRM:
+                    if(action.confirmed)
+                        database.deleteProject(action.uid)
                     break
             }
